requirements.txt
web3==6.11.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==7.4.3
eth-account==0.9.0
requests==2.31.0